feat: Refactor task dashboard by separating content into a new component and optimizing imports
